+++ test/pingdom_test.js
@@ -7,14 +7,14 @@ const server  = require('./helpers/server');
 const checksURL = 'http://localhost:3000/_healthchecks';
 
 
-describe('Pingdom', function() {
+describe('Pingdom run checks', function() {
 
   before(function(done) {
     server.ready(done);
   });
 
 
-  describe('runs checks, all healthy', function() {
+  describe('all healthy', function() {
 
     it('should receive response with status 200', function(done) {
       request(checksURL, function(error, response) {
@@ -25,7 +25,7 @@ describe('Pingdom', function() {
   });
 
 
-  describe('runs checks, URL not accessible', function() {
+  describe('URL not accessible', function() {
     before(function() {
       server.locals.error = true;
     });
@@ -43,7 +43,7 @@ describe('Pingdom', function() {
   });
 
 
-  describe('runs checks, response times out', function() {
+  describe('response times out', function() {
     before(function() {
       server.locals.timeout = ms('3s');
     });
@@ -63,7 +63,7 @@ describe('Pingdom', function() {
   });
 
 
-  describe('runs checks, response is 400', function() {
+  describe('response is 400', function() {
     before(function() {
       server.locals.status = 400;
     });
@@ -81,7 +81,7 @@ describe('Pingdom', function() {
   });
 
 
-  describe('runs checks, response missing expected content', function() {
+  describe('response missing expected content', function() {
     before(function() {
       server.locals.expected = 'Expected to see foo and also bar';
     });